% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_find.R
\name{lib_find}
\alias{lib_find}
\title{Find all packages used in scripts}
\usage{
lib_find(path = ".", verbose = FALSE)
}
\arguments{
\item{path}{Charater. Defines path to project file, a folder or an individual rmarkdown or script file. Defaults to current working directory}

\item{verbose}{Logical. Set as TRUE if results from all files are to be reported}
}
\value{
Returns a list containing all packages used in file(s) supplied by path
}
\description{
Find all packages used via library(), require() or with package::function() in your project, a folder, or in a file (.R, .Rmd, .Rmarkdown; not case sensitive).
}
\examples{
\dontrun{lib_find()}
}
