% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_interactive.R
\name{lib_interactive}
\alias{lib_interactive}
\title{Interactively choose which citations to keep}
\source{
relies on code in https://github.com/rstudio/DT/issues/93
}
\usage{
lib_interactive(libs,textformat = TRUE)
}
\arguments{
\item{libs}{Character (string) of package name(s) or Dataframe with a column "Package".}

\item{textformat}{Logical. Print as text version or as BibTeX.}
}
\description{
Opens an interactive shiny interface. Select desired packages, print them to console, copy them to the  clipboard and/or save them to disk (write new file or append). Note, that the download only captures entries visible in the table currently Incries number of rows shown to select from and write all packages to file.
}
\examples{
\dontrun{

lib_interactive(libs)
}
}
