% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_print.R
\name{lib_print}
\alias{lib_print}
\title{Print package citations}
\usage{
lib_print(libs, textformat = TRUE)
}
\arguments{
\item{libs}{Character (string) of package name(s) or Dataframe with a column "Package".}

\item{textformat}{Logical. Print as text version or as BibTeX.}
}
\value{
Returns a list of citations, either in text or BibTeX format.
}
\description{
Print package citations
}
\examples{
lib_print("base")
}
