% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_write.R
\name{lib_write}
\alias{lib_write}
\title{Write identified packages to a new (or append to existing) bibliography file}
\usage{
lib_write(libs, path_out, append, textformat = TRUE)
}
\arguments{
\item{libs}{Character (string) of package name(s) or Dataframe with a column "Package".}

\item{path_out}{Character. Path to new or existing bibliography file}

\item{append}{Logical. Determines if package bibliography should be appended or not}

\item{textformat}{Logical. Determines if plain-text or bibtex format should be used}
}
\description{
The bibliography of identified packages can be saved in two formats (plain-text or bibtex). Currently, the function does not recognize the text format from the file ending, so ensure to set the argument appropriately.
}
\examples{
\dontrun{

lib_write(libs, ".", append = FALSE)
}
}
