% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_interactive.R
\name{lib_interactive}
\alias{lib_interactive}
\title{Interactively choose which citations to keep}
\usage{
lib_interactive(libs, textformat = TRUE)
}
\arguments{
\item{libs}{ab}

\item{textformat}{ab}
}
\value{
Opens shiny
}
\description{
Interactively choose which citations to keep
}
\examples{
\dontrun{lib_interactive(libs)}
}
